Exposed config for EventStream + Updated plugin test
