+++ tests/plugins/event-log-sqlite.ts
@@ -1,20 +1,20 @@
-import type { DataStore } from "@tbd54566975/dwn-sdk-js";
-import { DataStoreSql } from "@tbd54566975/dwn-sql-store";
+import type { EventLog } from "@tbd54566975/dwn-sdk-js";
+import { EventLogSql } from "@tbd54566975/dwn-sql-store";
 import { getDialectFromUrl } from "../../src/storage.js";
 
 /**
- * An example of a plugin that is used for testing.
+ * An example of a plugin. Used for testing.
  * The points to note are:
  * - The class must be a default export.
  * - The constructor must not take any arguments.
  */
-export default class DataStoreSqlite extends DataStoreSql implements DataStore {
+export default class EventLogSqlite extends EventLogSql implements EventLog {
   constructor() {
     const sqliteDialect = getDialectFromUrl(new URL('sqlite://'));
     super(sqliteDialect);
 
     // NOTE: the following line is added purely to test the constructor invocation.
-    DataStoreSqlite.spyingTheConstructor();
+    EventLogSqlite.spyingTheConstructor();
   }
 
   /**
