+++ tests/scenarios/dynamic-plugin-loading.spec.ts
@@ -1,20 +1,15 @@
-import type { JsonRpcSuccessResponse } from '../../src/lib/json-rpc.js';
-import type { Readable } from 'readable-stream';
-
 import chaiAsPromised from 'chai-as-promised';
 import chai, { expect } from 'chai';
 import DataStoreSqlite from '../plugins/data-store-sqlite.js';
-import fetch from 'node-fetch';
+import EventLogSqlite from '../plugins/event-log-sqlite.js';
+import EventStreamInMemory from '../plugins/event-stream-in-memory.js';
 import sinon from 'sinon';
 
 import { config } from '../../src/config.js';
-import { createJsonRpcRequest } from '../../src/lib/json-rpc.js';
 import { DwnServer } from '../../src/dwn-server.js';
-import { getFileAsReadStream } from '../utils.js';
-import { v4 as uuidv4 } from 'uuid';
 
-import { Cid, DwnConstant, Jws, ProtocolsConfigure, RecordsRead, RecordsWrite, TestDataGenerator } from '@tbd54566975/dwn-sdk-js';
 import { DidDht, DidKey, UniversalResolver } from '@web5/dids';
+import CommonScenarioValidator from '../common-scenario-validator.js';
 
 // node.js 18 and earlier needs globalThis.crypto polyfill
 if (!globalThis.crypto) {
@@ -50,13 +45,21 @@ describe('Dynamic DWN plugin loading', function () {
 
     // NOTE: was not able to spy on constructor directly, so spying on a method that is called in the constructor
     const customDataStoreConstructorSpy = sinon.spy(DataStoreSqlite, 'spyingTheConstructor');
+    const customEventLogConstructorSpy = sinon.spy(EventLogSqlite, 'spyingTheConstructor');
+    const customEventStreamConstructorSpy = sinon.spy(EventStreamInMemory, 'spyingTheConstructor');
 
     // 1. Configure DWN to load a custom data store plugin.
     const dwnServerConfigCopy = { ...config }; // not touching the original config
+
+    // TODO: remove below after https://github.com/TBD54566975/dwn-server/issues/144 is resolved
+    // The default config is not reliable because other tests modify it.
     dwnServerConfigCopy.registrationStoreUrl = undefined; // allow all traffic
+
+    // dwnServerConfigCopy.messageStore = '../tests/plugins/message-store-sqlite.js'; // not working
     dwnServerConfigCopy.messageStore = 'sqlite://';
     dwnServerConfigCopy.dataStore = '../tests/plugins/data-store-sqlite.js';
-    dwnServerConfigCopy.eventLog = 'sqlite://';
+    dwnServerConfigCopy.eventLog = '../tests/plugins/event-log-sqlite.js';
+    dwnServerConfigCopy.eventStreamPluginPath = '../tests/plugins/event-stream-in-memory.js';
 
     // 2. Validate that the constructor of the plugin is called.
     // CRITICAL: We need to create a custom DID resolver that does not use a LevelDB based cache (which is the default cache used in `DWN`)
@@ -70,125 +73,11 @@ describe('Dynamic DWN plugin loading', function () {
     dwnServer = new DwnServer({ config: dwnServerConfigCopy, didResolver });
     await dwnServer.start();
     expect(customDataStoreConstructorSpy.calledOnce).to.be.true;
+    expect(customEventLogConstructorSpy.calledOnce).to.be.true;
+    expect(customEventStreamConstructorSpy.calledOnce).to.be.true;
 
     // 3. Validate that the DWN instance is using the custom data store plugin.
     const dwnUrl = `${dwnServerConfigCopy.baseUrl}:${dwnServerConfigCopy.port}`;
-    await sanityTestDwnReadWrite(dwnUrl);
+    await CommonScenarioValidator.sanityTestDwnReadWrite(dwnUrl);
   });
 });
-
-/**
- * Sanity test RecordsWrite and RecordsRead on the DWN instance.
- */
-async function sanityTestDwnReadWrite(dwnUrl: string): Promise<void> {
-  const alice = await TestDataGenerator.generateDidKeyPersona();
-  const aliceSigner = Jws.createSigner(alice);
-  // await registrationManager.recordTenantRegistration({ did: alice.did, termsOfServiceHash: registrationManager.getTermsOfServiceHash()});
-
-  // install minimal protocol on Alice's DWN
-  const protocolDefinition = {
-    protocol: 'http://minimal.xyz',
-    published: false,
-    types: {
-      foo: {}
-    },
-    structure: {
-      foo: {}
-    }
-  };
-
-  const protocolsConfig = await ProtocolsConfigure.create({
-    signer: aliceSigner,
-    definition: protocolDefinition
-  });
-
-  const protocolConfigureRequestId = uuidv4();
-  const protocolConfigureRequest = createJsonRpcRequest(protocolConfigureRequestId, 'dwn.processMessage', {
-    target: alice.did,
-    message: protocolsConfig.message,
-  });
-  const protocolConfigureResponse = await fetch(dwnUrl, {
-    method: 'POST',
-    headers: {
-      'dwn-request': JSON.stringify(protocolConfigureRequest),
-    }
-  });
-  const protocolConfigureResponseBody = await protocolConfigureResponse.json() as JsonRpcSuccessResponse;
-
-  expect(protocolConfigureResponse.status).to.equal(200);
-  expect(protocolConfigureResponseBody.result.reply.status.code).to.equal(202);
-
-  // Alice writing a file larger than max data size allowed to be encoded directly in the DWN Message Store.
-  const filePath = './fixtures/test.jpeg';
-  const {
-    cid: dataCid,
-    size: dataSize,
-    stream
-  } = await getFileAsReadStream(filePath);
-  expect(dataSize).to.be.greaterThan(DwnConstant.maxDataSizeAllowedToBeEncoded);
-
-  const recordsWrite = await RecordsWrite.create({
-    signer: aliceSigner,
-    dataFormat: 'image/jpeg',
-    dataCid,
-    dataSize
-  });
-
-  const recordsWriteRequestId = uuidv4();
-  const recordsWriteRequest = createJsonRpcRequest(recordsWriteRequestId, 'dwn.processMessage', {
-    target: alice.did,
-    message: recordsWrite.message,
-  });
-  const recordsWriteResponse = await fetch(dwnUrl, {
-    method: 'POST',
-    headers: {
-      'dwn-request': JSON.stringify(recordsWriteRequest),
-    },
-    body: stream
-  });
-  const recordsWriteResponseBody = await recordsWriteResponse.json() as JsonRpcSuccessResponse;
-
-  expect(recordsWriteResponse.status).to.equal(200);
-  expect(recordsWriteResponseBody.result.reply.status.code).to.equal(202);
-
-  // Alice reading the file back out.
-  const recordsRead = await RecordsRead.create({
-    signer: aliceSigner,
-    filter: {
-      recordId: recordsWrite.message.recordId,
-    },
-  });
-
-  const recordsReadRequestId = uuidv4();
-  const recordsReadRequest = createJsonRpcRequest(recordsReadRequestId, 'dwn.processMessage', {
-    target: alice.did,
-    message: recordsRead.message
-  });
-
-  const recordsReadResponse = await fetch(dwnUrl, {
-    method: 'POST',
-    headers: {
-      'dwn-request': JSON.stringify(recordsReadRequest),
-    },
-  });
-
-  expect(recordsReadResponse.status).to.equal(200);
-
-  const { headers } = recordsReadResponse;
-  const contentType = headers.get('content-type');
-  expect(contentType).to.not.be.undefined;
-  expect(contentType).to.equal('application/octet-stream');
-
-  const recordsReadDwnResponse = headers.get('dwn-response');
-  expect(recordsReadDwnResponse).to.not.be.undefined;
-
-  const recordsReadJsonRpcResponse = JSON.parse(recordsReadDwnResponse) as JsonRpcSuccessResponse;
-  expect(recordsReadJsonRpcResponse.id).to.equal(recordsReadRequestId);
-  expect(recordsReadJsonRpcResponse.error).to.not.exist;
-  expect(recordsReadJsonRpcResponse.result.reply.status.code).to.equal(200);
-  expect(recordsReadJsonRpcResponse.result.reply.record).to.exist;
-
-  // can't get response as stream from supertest :(
-  const cid = await Cid.computeDagPbCidFromStream(recordsReadResponse.body as Readable);
-  expect(cid).to.equal(dataCid);
-}
